requirements.txt

PyQt6
langchain
langchain-community
langchain-ollama
pypdf
pytesseract
Pillow
ollama